Try to mmap vdex file within the address range of the ELF file.
Add ELF section for the vdex file and mmap it there at runtime.
This ensures that the data is at predictable location, which
is needed to be able to reference it from native debug-info.

The vdex data is still always stored in the separate file,
and the ELF section is there only to reserve memory for it.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing

Change-Id: Ie2abd36c8b6617a527368e71f932998bbe5ad38c