Update patch set 3
Patch Set 3:

> > Nicolas, any more feedback? If not, I am going to submit....
 > 
 > I still find the whitelist of vector operations very brittle. But
 > won't object if you'd want to keep the performance numbers while
 > working on a proper fix.

I don't disagree :-)

If you look at the CanBeMoved in nodes_vector.h, it lists the following as false, for exactly the reason we are fixing now

HVecReplicateScalar
HVecExtractScalar
VecSetScalars
HVecStore             // cannot move for other obvious reason

So if the scheduler had respected this, we would not have had this issue, but I am guessing CanBeMoved is too restrictive for the scalar case.

Would you feel better with the very ugly 

(i->IsVecOperation() && !i->CanBeMoved())

so that at least we are using the logic that was put in noded_vector.h, or do you find this just as brittle, if not more?

Patch-set: 3
Reviewer: Gerrit User 1074526 <1074526@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0