Update 031-class-attributes for OpenJDK 9.
OpenJDK 9 javac no longer considers an anonymous class
declared inside a static method to be static. This CL
updates the test expectations accordingly.

Before this CL, the test only passes when compiled with
versions of OpenJDK < 9. After this CL, it passes on
OpenJDK 9 but fails on older versions.

See http://b/62290080 for context.

Bug: 69449021
Test: art/test/testrunner/run_build_test_target.py -j40 art-interpreter
Change-Id: I473451696304cfc9df256994042b0f93ada7d349