Update patch set 3
Patch Set 3:

> > Yes, I agree that's slightly better.
 > 
 > Artem, can you please make it so? I am also filing a feature
 > request on having SIMD types instead. Not sure about when we will
 > get to that, but good to track anyway.

Anyway, good news about SIMD types!

If we check CanBeMoved in IsSchedulingBarrier then it won't be possible to reorder independent HVecAdd and HVecStore (for which CanBeMoved returns false). Correct me please if I am wrong but to reorder those instruction is a valid thing to do? (two HVecStores are not permitted to be reordered due to side_effect dependencies)

This is exactly a problem that we don't want to use the same method for slightly different things without proper definition what the method returns.

Patch-set: 3
Reviewer: Gerrit User 1089808 <1089808@85c56323-6fa9-3386-8a01-6480fb634889>