[RFC] ART: Add support for VMStack.getAnnotatedStackTrace
Add Thread.CreateAnnotatedStackTrace to return an array that
contains a row for each stack frame with the StackTraceElement
describing the frame, an array containing all objects that are
locked at the described location, and optionally for the top
frame an object the thread is blocked on, waiting for or sleeping
on.

Add a test.

Bug: 70538431
Test: m test-art-host
Test: art/test/testrunner/testrunner.py -b --host -t 168
Change-Id: I0d92e3d8182c4a592549a6445854816f71afd29e