Update patch set 3
Patch Set 3:

I am really eager to do my best to make the situation less brittle however I'd like to better understand what you exactly mean. If someone introduces a new instruction it will be 'not schedulable' by default. So when taking a decision how a new instruction should behave in terms of scheduling one should investigate the restrictions, etc. For this ocasion it was my fault that I originally set the IsSchedulable=true for this VecReplicate. I think the reason for the situatuion to be 'brittle' is not with the scheduler but with the SIMD types (which is going to be changed someday in the future).

Lets look at the semantics (not the implementation and interfaces): an instruction have:

(target-independent behavior).
a) a property whether it can be moved during LICM/GVN

(target-dependent behavior).
b) a property whether it prevents the whole block to be scheduled.
c) a property whether no instruction should be reordered with this one.

Those property are not the same thing and can't be represented by a single function in my opinion. Then the question is whether they should be specified in the same place in instruction definition or scheduling part can live separately in scheduler.*. I can move IsSchedulable and IsSchedulingBarrier to the instruction declaration if you think it is more safe.

Patch-set: 3
Reviewer: Gerrit User 1089808 <1089808@85c56323-6fa9-3386-8a01-6480fb634889>