Load /data oat/odex files non-executable when only_use_system_oat_files is set.
Make it still ok to use the vdex and the class status in the oat file.
This means:
- If dex code is uncompressed in APK we'll use that
- If dex code is compressed in APK, we'll use the dex code in the vdex.

This is a temporary solution in order to almost mimic performance-wise what
full stack integrity wants to achieve (priv-apps running from verified,
uncompressed dex code in APK).

This is to assess the impact of full stack integrity on privileged apps.

This CL doesn't ensure any integrity, and doesn't check whether the vdex has
been "tainted".

bug: 30972906

Change-Id: If0a6ab26dd6211290db53b7f02d88d45aff68c0c