Update patch set 1
Patch Set 1:

Just as additional information: the other way of doing this is to always compile with 'javac -g' and then let the dexer create a build without locals information. With the DX command line that is using '--no-locals' for the dexer.

For D8 these two options are *almost* the same and will generate the same code in most cases. I believe there are still a few minor differences, but for the fuzzer I think either way is fine, so this is just FYI. :)

Patch-set: 1