Update patch set 1
Patch Set 1:

I've done some improvement to the algorithm which reduces the compile time increase from 12% to 0.8%. However it is still increase for a patch which doesn't bring significant performance improvements (on our benchmarks). So the question is: have you observed any performance impact of the patch on your benchmarks? If no, I will just implement a local version of the transformation for Loop peeling (that was the original intent).

Patch-set: 1
Reviewer: Gerrit User 1089808 <1089808@85c56323-6fa9-3386-8a01-6480fb634889>