Update patch set 4
Patch Set 4:

Well, in part that is already true.
The ELF section is really only used as pre-allocated space to be used for mmaping at run-time.  If we were to create odex file for dex-files, without haveing the vdex, it would work the same way.  The only question is: we might need a more generic name for the ELF section than ".vdex".

Patch-set: 4
Reviewer: Gerrit User 1065473 <1065473@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0