Fix dex2oat unquickening for update_input_vdex_ and reenable cdex tests
For the case where we are updating the input dex, avoid doing eager
unquickening. Eager unqucikening was causing the
Dex2oatUnquickenTest.UnquickenMultiDex test to fail because the
backing vdex is opened as a private map (writes don't propagate).

This change only unquickens eagerly if update_input_vdex is false
since this boolean already disabled dexlayout/cdex.

Test: test-art-host-gtest

Bug: 63756964
Bug: 70930171

Change-Id: Ib7b1e79157414677d288d217de3140f51171569c