Update patch set 1
Patch Set 1:

I need to look at compare_exchange a bit more carefully. Hopefully shortly.

All of this does raise the question of whether we actually want to maintain the runtime/atomic.h layer, or switch to using std::atomic directly. I think the advantages of keeping it are that it's less work in the short term, and it allows us to preserve the "JavaData" ordering, which doesn't really exist in C++, and which we may eventually want to approximate better than we currently do. The disadvantage is that it's another layer that anyone new to the project will have to understand. And aside from "JavaData", I think it's basically a "value subtracted" layer: The std::atomic interface is marginally better, since it allows abstraction with respect to memory ordering.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>