Use 28 bits for type check bit string.
And move ClassStatus to high bits of Class::status_.

Arbitrarily add the 4 extra bits to the third character of
the bit string. This CL is only about making those bits
available, the determination of how to use them for the best
impact (whether to have a 4th character or distribute them
differently among the three characters) shall be done later.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Test: Pixel 2 XL boots.
Test: testrunner.py --target --optimizing
Bug: 64692057
Change-Id: I38c59837e3df3accb813fb1e04dc42e9afcd2d73