Use 28 bits for type check bit string.
And reverse the order of fields in the Class::status_. This
avoids generated code size increase:
  - ClassStatus in high bits allows class initialization
    check using "status_high_byte < (kInitialized << 4)"
    which is unaffected by the low 4 bits of LHS instead of
    needing to extract the status bits,
  - the type check bit string in the bottom bits instead of
    somewehere in the middle allows the comparison on ARM
    to be done using SUB+LSLS rather than UBFX+CMP which can
    save 2 bytes (LSLS has a 16-bit variant, UBFX does not);
    it's also the same code size as with the old layout
    where we would have used LSR(S)+CMP.

Arbitrarily add the 4 extra bits to the third character of
the bit string. This CL is only about making those bits
available, the determination of how to use them for the best
impact (whether to have a 4th character or distribute them
differently among the three characters) shall be done later.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Test: Pixel 2 XL boots.
Test: testrunner.py --target --optimizing
Bug: 64692057
Change-Id: I38c59837e3df3accb813fb1e04dc42e9afcd2d73