Move quickening info logic to its own table
Added a table that is indexed by dex method index. To prevent size
overhead, there is only one slot for each 16 method indices. This
means there is up to 15 loop iterations to get the quickening info
for a method. The quickening infos are now prefixed by a leb
encoded length. This allows methods that aren't quickened to only
have 1.25 bytes of space overhead.

The value was picked arbitrarily, there is little advantage to
increasing the value since the table only takes 1 byte per 4 method
indices currently. JIT benchmarks do not regress with the change.

There is a net space saving from removing 8 bytes from each
quickening info since most scenarios have more quickened methods than
compiled methods.

For getting quick access to the table, a 4 byte preheader was added
to each dex in the vdex file

Removed logic that stored the quickening info in the CodeItem
debug_info_offset field.

The change adds a small quicken table for each method index, this
means that filters that don't quicken will have a slight increase in
size. The worst case scenario is compiling all the methods, this
results in 0.3% larger vdex for this case. The change also disables
deduping since the quicken infos need to be in dex method index
order.

For filters that don't compile most methods like quicken and
speed-profile, there is space savings. For quicken, the vdex is 2%
smaller.

Bug: 71605148
Bug: 63756964
Test: test-art-host

Change-Id: I89cb679538811369c36b6ac8c40ea93135f813cd