Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> (2 comments)
> 
> > Patch Set 8:
> > 
> > Alex, could you please paste the error you get with the native stubs? After discussing with Vladimir, it might just be a matter of tweaking a few checks.
> 
> I don't have it built right now but the failure was CHECK Failed (runtime->GetJit() != nullptr) https://android.googlesource.com/platform/art/+/master/runtime/stack.cc#814
> 
> This only happens when the jit is disabled. Otherwise it does actually all work.

Also I spent a few days trying to figure out if it was a case of tweaking a few checks/simple fixes and it seems that it will take a much deeper fix to make it all work. (Although I admit I'm not the most familiar with this code).

Patch-set: 8
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0