Update patch set 8
Patch Set 8:

> [...] the failure was CHECK Failed
 > (runtime->GetJit() != nullptr) https://android.googlesource.com/platform/art/+/master/runtime/stack.cc#814
 > 
 > This only happens when the jit is disabled. Otherwise it does
 > actually all work.

That would mean that the method on the stack was called with an AOT-compiled stub but method->GetOatMethodQuickCode(class_linker->GetImagePointerSize()) cannot find it anymore. I would think that once we actually point a method to an AOT compiled stub, we have already verified that it can be executed and that determination cannot change.

Patch-set: 8
Reviewer: Gerrit User 1018108 <1018108@85c56323-6fa9-3386-8a01-6480fb634889>