Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > [...] the failure was CHECK Failed
>  > (runtime->GetJit() != nullptr) https://android.googlesource.com/platform/art/+/master/runtime/stack.cc#814
>  > 
>  > This only happens when the jit is disabled. Otherwise it does
>  > actually all work.
> 
> That would mean that the method on the stack was called with an AOT-compiled stub but method->GetOatMethodQuickCode(class_linker->GetImagePointerSize()) cannot find it anymore. I would think that once we actually point a method to an AOT compiled stub, we have already verified that it can be executed and that determination cannot change.

Yup. You see the problem. Since we are so early we know it's only a few gc related frames on the stack which are all fine but we cannot get rid of the native transitions because they are already there as the gc-threads 'wait' functions.

Changing loading to take place earlier might be a possible solution but that would be rather complicated too (unless we just special-cased some plugins).

Patch-set: 8
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0