Update patch set 5
Patch Set 5:

(9 comments)

> (9 comments)
 > 
 > The runtime code seem correct, I'm just not sure about how this
 > will affect the compiled code size?

It does not affect the current generated code size or future expected code size for ARM64. ARM may be affected but, depending on which cases are more common, it may be a gain (no CMP equivalent of SUB(immediate,T4) for 9-12 bit #imm) or a loss (no 16-bit SUB equivalent of CMP(register,T2)).

Patch-set: 5
Label: Presubmit-Ready=0