Move debug info offsets into a side table
Add a compact side table for figuring out the debug info offsets
for a given method index. This reduces dex size by ~1.2%.

The debug table is keyed by method index and has leb encoded
offsets for the offsets. This means the table is smaller if debug
infos are encoded by method index order.

To prevent expansion for method indicies without debug info, there
is a bitmap that specifies if a method index has a debug info offset.

Motivation: Reduce code item size and allow more deduping in the
future.

Test: test-art-host
Bug: 63756964

Change-Id: Ib983e85c1727f58c97676bde275f4a9756314da0