Update patch set 13
Patch Set 13:

Hi folks

This is still work in progress, but the bottom layer now covers the complete set of VarHandle kinds in var_handle.{h,cc}.

There's nearly 2000 additional lines in var_handle.cc in the PS today. This is unwieldly. There seem to be two ways of reducing the current approach:

(1) implement all accessors with a strong memory ordering. On interpreter timescales this may not have any noticeable effect.

(2) there is specialization over signed and unsigned types. With the exception of getAndAdd style accessors, the code could treat signed and unsigned as the same. This is perhaps less intuitive to the reader.

Does this seem reasonable? If you any thoughts on beating the size down further or a better factoring, suggestions are definitely welcome!

The code also adds 5 fields from ByteBuffer to WellKnownClasses. Perhaps this should be a mirror class?

Thanks
Orion

Patch-set: 13
CC: Gerrit User 1014443 <1014443@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>