Compiler changes for bitstring based type checks.
We guard the use of this feature with a compile-time flag,
set to true in this CL.

Boot image size for aosp_taimen-userdebug in AOSP master:
  - before:
    arm boot*.oat: 63604740
    arm64 boot*.oat: 74237864
  - after:
    arm boot*.oat: 63531172 (-72KiB, -0.1%)
    arm64 boot*.oat: 74135008 (-100KiB, -0.1%)

The new TypeCheckBenchmark yields the following changes:
                               32-bit        64-bit
  timeCheckCastLevel1ToLevel1   9.32->13.63   9.33->13.65
  timeCheckCastLevel2ToLevel1  12.91->13.63  12.92->13.65
  timeCheckCastLevel3ToLevel1  18.04->13.63  19.43->13.62
  timeCheckCastLevel9ToLevel1  40.45->13.64  40.49->13.62
  timeCheckCastLevel9ToLevel2  37.47->12.20  37.47->12.20
  timeInstanceOfLevel1ToLevel1 11.50->15.78  11.48->15.80
  timeInstanceOfLevel2ToLevel1 15.27->15.77  15.28->15.81
  timeInstanceOfLevel3ToLevel1 20.08->15.80  20.51->15.77
  timeInstanceOfLevel9ToLevel1 45.19->15.80  45.10->15.79
  timeInstanceOfLevel9ToLevel2 42.19->14.37  42.23->14.39
This suggests that the bitstring typecheck should not be
used for exact type checks which would be equivalent to the
"Level1ToLevel1" benchmark. Whether the implementation is
a beneficial replacement for the kClassHierarchyCheck and
kAbstractClassCheck on average depends on how many levels
from the target class (or Object for a negative result) is
a typical object's class.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing --jit
Test: testrunner.py --host -t 670-bitstring-type-check
Test: Pixel 2 XL boots.
Test: testrunner.py --target --optimizing --jit
Test: testrunner.py --target -t 670-bitstring-type-check
Bug: 64692057
Bug: 71853552
Bug: 26687569
Change-Id: I538d7e036b5a8ae2cc3fe77662a5903d74854562