Use efficient encoding for compact dex code item fields
Use a format for compact dex code items. In the common case, there
is only 4 bytes worth of fields.
4 bits: register size, ins size, outs size, tries size
11 bits: insns size
5 bits: if there is a preheader extension for each field.

Preheader may use up to 12 extra bytes. The preheader non empty for
~1% of code items. The current preheader encoding is suboptimal size
wise but is meant to be fast and easy to decode.

Reduces vdex size by an average of 4.3% on golem benchmarks.

TODO: Measure interpreter performance.

Test: test-art-host

Bug: 63756964
Change-Id: I836fd098063dd1573a3dfb6ac465a6d33d2cb2c9