Use efficient encoding for compact dex code item fields
Use a compact format for compact dex code items. In the common case,
there are only 4 bytes worth of fields.
4 bits each: register size, ins size, outs size, tries size
11 bits: insns size
5 bits: if there is a preheader extension for each field.

There is an optional preheader before each code item that describes
fields that don't fit in the fast path case. The preheader may use up
to 12 extra bytes. Experiments showed that the preheader is only non
empty for ~1% of code items.

Changed the code item alignment to be 2 byte aligned for code items
that don't have payload instructions that require 4 byte alignments.

Results:
The CL reduces vdex size by an average of 4.3% on golem benchmarks.
Interpreter performance does not signficantly change on golem.

Test: test-art-host

Bug: 63756964

Change-Id: I836fd098063dd1573a3dfb6ac465a6d33d2cb2c9