Add `is_boot_class_path` flag to DexFileLoader/DexFileVerifier
This will be required by CL Ifc237ff8a35a8b470b7fc682a9cb879370d1e6e9
but is submitted seperatly as it touches quite a few files.

The additional parameter on DexFileLoader entry points stipulates
whether the loaded dex file belongs to boot class path and should
be dex-verified as such. This is needed because the encoding of
access flags introduced in the other CL would fail a check in
DexFileVerifier whether at most one of the three visibility flags
is set.

Test: m test-art-host
Bug: 64382372
Change-Id: I1bee6f639e961114e4a259dc499aca0036fe58a0