Revert^2 "Ensure that methods requiring interpreter entrypoint always have it."
The code in test 067-preemptive-unpark is able to cause a method to
have a null entrypoint. This is a valid (though rare) state. The check
we had did not account for it and failed. Made the DCHECK explicitly
allow for null entrypoints.

This reverts commit c971f6e06a43e70a2d86cb4b3ad0e13a9ae9a06c.

Reason for revert: fixed issue causing test 067 to fail.

Test: ./test.py --host -j50
Test: ./test/testrunner/testrunner.py -j50 -t 067-preemptive-unpark

Bug: 62821960

Change-Id: I537e05fd52fa4c8859d4102ac9475cacd75d7437