Generate mini-debug-info for interpreted methods as well.
Separate CL as this affects the size of the boot image.
(+1.5%/170k to the size of boot.oat on arm on aosp)

This will allow resolving interpreted methods in libunwind.

This is the simplest solution. I intend to revert this CL when
we teach libunwind to understand the dex file format directly.

Bug: 71579677
Test: m test-art-host-gtest
Test: testrunner.py --host -t 137
Change-Id: I290f31ccacaf868e3e76a3d3574341c709cb82fe