Update patch set 3
Patch Set 3: Presubmit-Ready+1

It turns out my previous attempt to fix the flakiness didn't handle all of the cases. If concurrent GC was triggered near the end of the loop, it could get stuck behind the explicit GC and end up starting for real just before the round of calls to register native allocation. Because GC started before the dummy object was allocated, it would not collect the object, but by the time GC had completed, we had finished all the calls to register native allocation and no new GC was triggered.

I switch to a different approach that, fingers crossed, will not be flaky. I was not able to reproduce failures with this new approach that were relatively easy to reproduce with the previous approach when running the test in a loop on a machine with high load.

I still need to check the renderscript cts test.

Patch-set: 3
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1