Restore bitmap marking in non-moving space in ConcurrentCopying::Copy.
When when copying an (evac) from-space reference to the non-moving
space in ConcurrentCopying::Copy, also mark it in the heap mark
bitmap, like it used to be. (This marking operation was removed in
https://android-review.googlesource.com/257901.)

This addresses an issue with Sticky-Bit (Generational) CC collection,
where an assertion in
ConcurrentCopying::AssertToSpaceInvariantInNonMovingSpace could fail
when checking a reference in the non-moving space, that would happen
to be unmarked (and not on the allocation stack).

Test: ART run-tests & gtests, libcore tests, JDWP tests (host & device)
Test: Device/emulator boot test
Bug: 67628039
Bug: 12687968
Change-Id: Id21540a77c6c9f90e819377473951b4887f2f6bf