Move elf_builder.h and stream helpers to the runtime
Move them since I will need to create an in-memory ELF file at runtime
to communicate with libunwind (no files are actually written to disk).

This will, in turn, allow libunwind to unwind through interpreted code.
This uses the existing ELF-based protocol instead of inventing new one.

This code path does not depends on the compiler, and it would, indeed,
be particularly useful on a platform without the compiler library.

Test: m checkbuild
Change-Id: Ib466f73582220710c1e8fb9d051ba602d4b625e1