Update patch set 1
Patch Set 1:

Are you saying that you're currently experiencing this error, or that you used to? It should already be fixed by CLs [1-2]. We already have tests that would fail with the error; do we really need this additional one?

For context: The error message indicates that the .java code in question was compiled against OpenJDK 9 APIs but run against Android APIs. This is because the return type of ByteBuffer.limit() is ByteBuffer in OpenJDK 9, but it is Buffer in OpenJDK 8 and in Android.

[1] https://android-review.googlesource.com/c/platform/build/+/573902
[2] https://android-review.googlesource.com/c/platform/build/soong/+/578623

Patch-set: 1
Reviewer: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>