Update patch set 1
Patch Set 1: Code-Review-1

I think a regression test, if any, would really belong in 070-nio-buffer, not here.

That test does not currently test for a regression because we worked around the issue there in:

https://android-review.googlesource.com/c/platform/art/+/570282

I don't know of any reason not to revert that CL now. So I'll create a revert of that CL and test it; it'd probably be better to submit such a revert than to submit this CL.

Patch-set: 1
Reviewer: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1