Update patch set 1
Patch Set 1:

Well, I did initially consider it as well, but after a bit of though it does not really feel like compiler.  It is ELF utility code, and we already have plenty of that in the run-time.  If anything it moves the related code together.

I could of course keep this code in the compiler, and do the dlopen/dlsym dance to get to it, but it feels like an overkill for utility code.  I did briefly discuss those two options with Nicolas this morning and he was learning towards moving the utility code to run-time.  He might change might after seeing this CL though.

Personally, I would like to stick to this approach since it works without if the compiler library is missing from the device. In fact, that is the scenario when this code would be most important.  I would say that, the fact this code is most needed when there is no compiler, suggests this should not belong to the compiler.

Patch-set: 1
Reviewer: Gerrit User 1065473 <1065473@85c56323-6fa9-3386-8a01-6480fb634889>