Update patch set 5
Patch Set 5:

(8 comments)

@Andreas: I want your opinion on how this should be benchmarked.

IIUC, Golem only accepts art/ patches, so if we want to set certain methods as hidden, we'd need to merge a CL into the repo. I'd like to avoid doing that, which rules out seeing the effects of soft failed classes. On the other hand, I seriously doubt that our benchmarks statically link against @hide class members, so there should not be any anyway.

Similarly, we'd have to merge a CL to see the effects of printing a warning. It's a slow path, don't care.

So that leaves us with these checks which should always return true or not print a warning. Is that okay with you?

Patch-set: 5
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>