Update patch set 1
Patch Set 1:

Yes, global system property, which will be set in the build.
(set to true on the devices where we can afford it)

Alternatively, I would like it to default to the build flag which controls mini-debug-info for boot.oat.  Either by making the C++ code default to it, or possibly better, propagating in the build.

Patch-set: 1
Reviewer: Gerrit User 1065473 <1065473@85c56323-6fa9-3386-8a01-6480fb634889>