Update patch set 7
Patch Set 7:

Do we get the extra heap space back in the end, after the GC?

Do we ensure that we don't do this for the CMS collector? Or we don't care because nobody uses it anymore? Tripling the startup heap size with CMS is probably not good.

Patch-set: 7
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>