Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Do we get the extra heap space back in the end, after the GC?
> 
> Do we ensure that we don't do this for the CMS collector? Or we don't care because nobody uses it anymore? Tripling the startup heap size with CMS is probably not good.

The duration of the CameraLaunch is around 400ms, less than the 2s window chosen by us, so the GC doesn't even get triggered. Once the GC is triggered, I don't see any reason why we won't get the space back. In fact, all the regions that will be allocated during the window will be candidates for evacuation, and hence we'll actually retrieve back space for future allocations.

IMHO we shouldn't generalize that this approach leads to triple heap consumption. The number is so small that tripling isn't a big deal. Basically the only downside of this approach is that we don't get to collect garbage before 2s, unless the app allocates so fast that it manages to reach half of the heap size before that.

Patch-set: 7
Reviewer: Gerrit User 1258954 <1258954@85c56323-6fa9-3386-8a01-6480fb634889>