ahat: Fix make dependency for $(AHAT_TEST_DUMP_PROGUARD_MAP)
The make rule specified that it depended on $(proguard_dictionary),
but that is only a file path and not the name of a rule for making
that file. This caused the build to break when $(proguard_dictionary)
didn't happen to be created by the time the rule was being run.

The file is created as a side-effect of making ahat-test-dump;
therefore, this CL changed the rule to say that it depended on
that rule being made first.

Bug: 71710349
Test: Treehugger

Change-Id: Ieba37296646359bd5af4e66fe2ac70702906fff1