Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> IMHO we shouldn't generalize that this approach leads to triple heap consumption. The number is so small that tripling isn't a big deal. Basically the only downside of this approach is that we don't get to collect garbage before 2s, unless the app allocates so fast that it manages to reach half of the heap size before that.

I wouldn't have classified 4MB of additional dirty memory per process as small. We may have less than a GB and a large number of processes. In the absence of more measurements, I'd guess that this is too much to make this profitable, unless we manage to madvise it away at the next GC, which is what I would expect for CC, but not CMS.

It might be that this 4MB difference is unique to Camera, in which case it becomes less of a concern.

Patch-set: 7
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>