Update patch set 8
Patch Set 8:

Is this really the right tradeoff ? +4MB of peak heap (which is a lot, quite honestly) for 10ms of startup time doesn't seem worth it to me. I won't object if this has been discussed and is a conscious decision. I would also like to see more detailed measurements (not just for one app).

Patch-set: 8
CC: Gerrit User 1014443 <1014443@85c56323-6fa9-3386-8a01-6480fb634889>