Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> So you're probably missing out on RTP not being run in certain cases after inlining. RTP is supposed to do that non-null refinement for you.
> 
> I think instead, you could detect the cases where running RTP would be beneficial after inlining (right  now it's only if the return type is more specific).
> 
> As for the changes in instruction_simplifier, all these Dominates call make me worry of the compile time costs.

We need this optimization after DCE, since that "cuts" the always throws edges from the CFG, and thus creates the new opportunities for this elimination.

Note that the dominates-check is only done after a very specific pattern for if (object != null), so I doubt it will be called much (worry I). Or perhaps you are worried about the already existing dominates-check in CanEnsureNotNullAt() (worry II), which is not called more often from VisitNullCheck()?

This seemed a clean and elegant place to add the optimization and, as the checker test reveals, it even cleaned up a few other cases that were outside the immediate scope of the Kotlin targeted optimization.

But, if you are are still worried, I can offer two alternatives

(1) Don't add HasNonNullControlDependence() to CanEnsureNotNullAt() and call CanEnsureNotNullAt() from VisitNullCheck, but just call  HasNonNullControlDependence() from m VisitNullCheck directly. This solves worry (II).

(2) Or, to solve both worries, I could move this new optimization into DCE and only test if statements that were just optimized with the always-throws technique. This ensures that the dominator tests are practically never called, and in the rare cases they are called, only were they could be useful.

How does that sound?

Patch-set: 1
Reviewer: Gerrit User 1074526 <1074526@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0