Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Thanks. I can reproduce the failure using your steps. I've filed an
>  > internal bug for this (bug 72491947). I'll look into it over the
>  > next days; I'll have to figure out what script calls javac with the
>  > wrong bootclasspath in this situation. How badly are you blocked by
>  > this issue?
> 
> One of the benchmarks in our CI crashes because of that.
> 
> P.S.: When developing our wrapper scripts for testing we followed the similar approach to one here: https://chromium.googlesource.com/chromium/tools/build/+/master/scripts/slave/recipes/art.py. I wonder whether chromium will also have this issue.

Thanks again for the reproduction steps. I've dug into this now, and the reason why 162-method-resolution/src/Main.java fails is that the test runners build code under src/ against the current default host-side toolchain; therefore, they do not support calling API from code below src/ where that API has changed in an incompatible way in the current default version of the host-side toolchain (currently OpenJDK 9). In this particular case, OpenJDK 9, ByteBuffer.limit() etc.'s return type is ByteBuffer, whereas in OpenJDK 8 and in Android it is Buffer. Hence, calling such API from .java sources in art/test/.../src is currently not supported.

If we wanted to do this from an ART test, the (currently) most appropriate fix would be to rename src/ to art-src/, which will cause the test runner to build the code against the Android API. From that perspective, the code in this CL is expected to fail.

You said that the original reason why you looked into this was because "one of the benchmarks in [your] CI crashes because of that". The error is likely in whatever scripts / javac commands are used to build that benchmark (the javac invocation needs to be passed the correct API stubs via --bootclasspath or similar, or at least --release 8). Are you saying that your benchmark is being built via art/test.py?

Patch-set: 1
Reviewer: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>