Visit targets of proxy methods when visiting thread roots.
The target of a non-static proxy method (`this` object), stored in the
proxy method's stack frame, needs to be visited as GC root. This is
especially important in the case of a moving GC, where the proxy
instance may be moved like any object.

Fix initially provided by Robert Vollmer.

Test: m test-art-host
Bug: 70216372
Change-Id: Iea27a8eba51ccd9c9055efaf6b263892830170b5