Change sleep to busy loop.
Using a sleep to pause the thread to unwind can, every once in a while,
wind up at an instruction for which there is no unwind information.
Instead, use a busy loop that stays in the test code to avoid this problem.

Also, modify the java read code slightly to make it easier to tell when
the remote process to be unwind is misbehaving.

Bug: 73048324

Test: Ran cfi tests.
Change-Id: I3aa084495c9278b133ed6dbfda6373063db365af