Update patch set 1
Patch Set 1:

This change is intended to catch a specific situation, when the following sequence of events takes place:

An AbstractClass is defined by ClassLoader1.
A ConcreteClass extends AbstractClass is defined by ClassLoader2.
ClassLoader2 eventually gets unloaded, its LinearAlloc with all the ArtMethods of ConcreteClass is marked as "to be reused".
The LinearAlloc memory is reused, its entire content is zeroed.
A method, that have calls to a AbstractClass'es method, is being JITed.
Then, HInliner tries to get SingleImplementation code for that method. Since it's an ArtMethod that was already erased, a crash/unpredicted behaviour happens.

-----

The patch fixes it in a straightforward way - simply check everything that is being deleted is see if could be a SingleImplementation for something, then update that SingleImplementation field.

If it looks suboptimal because of the time complexity, I guess there's also an option to have some sort of an unordered map SingleImplementations <-> LinearAllocs to fasten the traversal. Obviously it will cost some KBs.

Patch-set: 1
Reviewer: Gerrit User 1121591 <1121591@85c56323-6fa9-3386-8a01-6480fb634889>